Use int instead of uint
While using uint might be technically more correct, it led to worse code
as it still needed to be converted back to an int in various places, for
example when compared to len().

+++ src/controllers/notificationController/list.go
@@ -10,7 +10,7 @@ import (
 )
 
 func List(c *fiber.Ctx) error {
-	pageNumber, ok := pager.GetCurrentPageNumber(uint(c.QueryInt(pager.Key, 1)))
+	pageNumber, ok := pager.GetCurrentPageNumber(c.QueryInt(pager.Key, 1))
 
 	if !ok {
 		return c.SendStatus(404)
